emr-ai-project 1/requirements.txt
ultralytics
pillow
numpy
opencv-python
matplotlib